+++ deployment/updates/scripts/add_permissions/2016_10_18_update_play_server_list_flavor_asset_service_permission.php
@@ -1,0 +1,9 @@
+<?php
+/**
+ * @package deployment
+ * @subpackage kajam.roles_and_permissions
+ */
+
+$addPermissionsAndItemsScript = realpath(dirname(__FILE__) . '/../../../../') . '/alpha/scripts/utils/permissions/addPermissionsAndItems.php';
+$config = realpath(dirname(__FILE__)) . '/../../../permissions/service.flavorasset.ini';
+passthru("php $addPermissionsAndItemsScript $config");