+++ release-notes.md
@@ -10,6 +10,7 @@
 
 ### Deployment scripts ###
  - Reindex sphinx and re-populate schedule-event
+ - Run 'php deployment/updates/scripts/add_permissions/2016_10_18_update_play_server_list_flavor_asset_service_permission.php'
 #### Known Issues & Limitations ####
 None.